# viewmodels/auth.go
package viewmodels

type Token struct {
	Token string `json:"token" example:"eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJuYW1lIjoiSm9obiBEb2UifQ.b-bQWDLukCsP-kz255shgiDNoPjycxSizc1cIljoEic"`
} // @name Token